Move about content to info command instead
